+++ src/config/validate-config.ts
@@ -1,12 +1,12 @@
 import { Config } from "./config.js";
-import { ValidateBundle } from "./validate-bundle.js";
+import { validateBundle } from "./validate-bundle.js";
 
 /**
  * Throws an exception if the provided raw config contains invalid data.
  * @param config - Raw configuration to validate.
  * @public
  */
-export function ValidateConfig(config: Config): void {
+export function validateConfig(config: Config): void {
     // If bundle key exists, value must be an object
     if ("bundle" in config) {
         const bundles = config.bundle;
@@ -17,7 +17,7 @@ export function ValidateConfig(config: Config): void {
         else {
             // Each property must be an object (for owned properties)
             for (const bundleName of Object.getOwnPropertyNames(bundles)) {
-                ValidateBundle(bundles[bundleName], bundleName);
+                validateBundle(bundles[bundleName], bundleName);
             }
         }
     }